requirements.txt
bandit
python-dotenv
numpy
pandas
pytest
matplotlib
sqlalchemy
psycopg2
sphinx